Reorganize utils
